Revert "drm/bridge: sn65dsi83: Register and attach our DSI device at probe"
The MIPI-DSI host registers itself and then tries to find the MIPI-DSI device
and vice-versa. This ends up in a deadlock, with both drivers deferring
infinitely. Move the MIPI-DSI device registration back to the attach function
to sync with the MIPI-DSI host registration to resolve this.

This reverts commit 6ef7ee48765fa3067858d11ecdf3acbc7c19df80.

This patch is based on 7283e6fcb16985b22519eb4dc8d6ac9984e8cb23

Co-authored-by: Alexander Herfurtner <alexander.h@variscite.com>
Signed-off-by: Luan Rafael Carneiro <luan.r@variscite.com>