Revert "drm/bridge: ti-sn65dsi83: Fix enable/disable flow to meet spec"
The proposed change aimed to correct the enable/disable flow of the
ti-sn65dsi83 controller to originally follow the sequence specified in the
datasheet. This involved splitting the enable() method into two parts:
pre_enable(), responsible for steps 2-5 (register initialization,
PLL enablement, etc.), and enable(), which would cover the final step of
error status verification.

However, after applying this change, a problem related to the PLL
(Phase-Locked Loop) was identified. The PLL error occurs because the separation
of the steps seems to affect the timing of enablement and initialization. The
change in execution order and the moment when the PLL is enabled, particularly
moving this responsibility to pre_enable(), is not compatible with the hardware
requirements.

This reverts commit dd9e329af7236e34c566d3705ea32a63069b9b13.

Signed-off-by: Luan Rafael Carneiro <luan.r@variscite.com>