ARM: dts: symphony: fix watchdog pin
Fix pin to use GPIO1_IO08. This will result in a working watchdog-reboot
when using the circuitry described in the schematic.

GPIO1_IO01 is going to an unmounted header, so it is not easily possible
to connect to the external symphony board reset and it is also not
documented in the Schematic as "the way to connect the Watchdog".

Signed-off-by: Matthias Proske <matthias.p@variscite.com>